+++ praticas/pratica01/caixa_eletronico.h
@@ -1,0 +1,6 @@
+#ifndef CAIXA_ELETRONICO_H
+#define CAIXA_ELETRONICO_H
+
+void cQtdNotas(int valor);
+
+#endif
