+++ praticas/pratica01/imc.h
@@ -1,0 +1,6 @@
+#ifndef IMC_H
+#define IMC_H
+
+void calcularIMC (float peso, float altura);
+
+#endif
