+++ praticas/pratica01/media.h
@@ -1,0 +1,6 @@
+#ifndef MEDIA_ARITMETICA_H
+#define MEDIA_ARITMETICA_H
+
+float calcularMedia(int n);
+
+#endif
