+++ praticas/pratica01/potencia.h
@@ -1,0 +1,6 @@
+#ifndef POTENCIA_H
+#define POTENCIA_H
+
+int calcularPotencia(int base, int expoente);
+
+#endif
